requirements.txt
python-barcode==0.10.0
treepoem==3.3.1
